Move DTOs in dto subfolders and remove @NonNullByDefault annotations from them
